fix(user.service.ts): :bug: fix type mismatch for dates in user creation service
Fixed a type mismatch on the "createdAt" parameter of a user between the database and the service. It was merely an internal issue and should not affect the endpoint, however a confirmation code line was also added to the service to reflect another mismatch between the database and the service treated in a previous commit.