refactor: :recycle: change interface and function names
Changed the interfaces present on the models to a cleaner style, using the Partial class. Changed the functions 'succeeded' and 'failed' to infinitive tense, to follow the international standard better and reduce line length. Moved environment variable calls from the top of a few modules to inside some functions, so as to make them load only after said variables have been properly accounted for and loaded, on the config file. Added native mongoose timestamps to models.